Changed the PkgConfigDeps structure. Fixed bad imports. Skipping questionable test
