--- conan/tools/gnu/pkgconfigdeps/__init__.py
@@ -1,1 +1,0 @@
-from conan.tools.gnu.pkgconfigdeps.pkgconfigdeps import PkgConfigDeps
