+++ conans/client/generators/markdown.py
@@ -9,7 +9,7 @@ from conan.tools.cmake.cmakedeps.cmakedeps import CMakeDeps
 from conan.tools.cmake.cmakedeps.templates import (
     CMakeDepsFileTemplate,
     get_cmake_package_name as cmake_get_file_name)
-from conan.tools.gnu.pkgconfigdeps.pc_info_loader import (
+from conan.tools.gnu.pkgconfigdeps import (
     _get_component_name as pkgconfig_get_component_name,
     _get_name_with_namespace as pkgconfig_get_name_with_namespace,
     _get_package_name as pkgconfig_get_package_name
